Authorisations without policy

+++ app/Http/Controllers/ThreadsController.php
@@ -120,6 +120,16 @@ class ThreadsController extends Controller
     public function destroy($channel, Thread $thread)
     {
         //
+
+        if($thread->user_id != Auth()->id()){
+
+            if(request()->wantsJson()){
+                return response(["status"=>"Permision Denied"],403);
+            }
+
+            return redirect("/login");
+        }
+
         $thread->delete();
